Never check process environment in stand-alone mode. (#13)
As pointed out by @shomikverma, this check gets in the way when
setting `DYLD_FALLBACK_LIBRARY_PATH` instead of `DYLD_LIBRARY_PATH`,
which may be the better setup. It is the user's reponsibiliy to
correctly configure the environment, we should not interfere with
that as long as we're unsure if we have the best solution. The
sanity check of environment variables was therefore removed.