Add support for material property groups. (#90)
As pointed out in issue #78, the material property groups, such as
"Basic" for the material "medium 1" in the demo model, did not show up
in the model tree as displayed by `mph.tree()`. We also could not
access those properties via the `Node.property()` method.

We fix that by also considering "property groups" (in addition to
"feature groups") as Java containers for children of a given node
when traversing the model tree.