Removed Trove classifiers for Python versions.
This information is redundant and easily outdated in the project
metadata. The `requires-python` field in the `[project]` section
of `pyproject.toml` is intended to be the single source of truth.
And since MPh is a pure-Python library, the supported Python
version ultimately depends on what's supported by its dependencies
with native code components, namely JPype (first and foremost)
and NumPy (less critically).